+++ helpers/helpers.php
@@ -31,7 +31,7 @@ function stm_wpcfto_nonces() {
 		'wpcfto_save_settings',
 		'get_image_url',
 		'wpcfto_upload_file',
-		'wpcfto_search_posts'
+		'wpcfto_search_posts',
 	);
 
 	$nonces_list = array();
@@ -51,11 +51,14 @@ add_action( 'admin_head', 'stm_wpcfto_nonces' );
 add_action( 'wp_head', 'stm_wpcfto_nonces' );
 
 
-add_action( 'wp_ajax_stm_wpcfto_get_settings', function() {
-	$source = ( isset( $_GET['source'] ) ) ? sanitize_text_field( $_GET['source'] ) : '';
-	$name   = ( isset( $_GET['name'] ) ) ? sanitize_text_field( $_GET['name'] ) : '';
-	wp_send_json(wpcfto_get_settings_map($source, $name));
-});
+add_action(
+	'wp_ajax_stm_wpcfto_get_settings',
+	function() {
+		$source = ( isset( $_GET['source'] ) ) ? sanitize_text_field( $_GET['source'] ) : '';
+		$name   = ( isset( $_GET['name'] ) ) ? sanitize_text_field( $_GET['name'] ) : '';
+		wp_send_json( wpcfto_get_settings_map( $source, $name ) );
+	}
+);
 
 function wpcfto_get_settings_map( $source, $name ) {
 	if ( 'settings' === $source ) {
@@ -73,6 +76,8 @@ function wpcfto_get_settings_map( $source, $name ) {
 
 			$settings = $option_page['fields'];
 		}
+
+		$settings_data = apply_filters( 'filter_settings_data_values', $settings_data );
 
 		foreach ( $settings as $section_name => $section ) {
 			foreach ( $section['fields'] as $field_name => $field ) {