+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 4.2.9
+- **Fix:** Added filter for data settings.
+## 4.2.8
+- **Fix:** Button List field improvement.
 ## 4.2.7
 - **Fix:** Strings translation issue.
 ## 4.2.6